+++ src/main/java/com/uooconnect/main/FrameworkTest.java
@@ -1,0 +1,12 @@
+package com.uooconnect.main;
+
+import com.uooconnect.util.BrowserStart;
+
+/**
+ * this is a test main class, don't use in other applications
+ */
+public class FrameworkTest {
+    public static void main(String[] args) throws InterruptedException {
+        BrowserStart.invokeBrowser();
+    }
+}
